+++ src/gll_library.cpp
@@ -9,20 +9,20 @@ gll_library::gll::gll()
 gll_library::gll::gll(const double alpha, const double beta, const int ngll)
     : alpha(alpha), beta(beta), ngll(ngll){}
 
-float gll_library::pnleg(const float z, const int n){
+double gll_library::pnleg(const double z, const int n){
     // Generate Lagendre polynomials using recurrance relation
     // (l+1)P_(l+1)(x)-(2l+1)xP_l(x)+lP_(l-1)(x)=0 
 
     if (n==0) throw std::invalid_argument("value of n > 0");
     
-    float p1, p2, p3, double_k;
+    double p1, p2, p3, double_k;
 
     p1 = 1.0;
     p2 = z;
     p3 = p2;
 
     for (int k = 1; k < n; k++){
-        double_k = static_cast<float>(k);
+        double_k = static_cast<double>(k);
         p3  = ((2.0*double_k+1.0)*z*p2 - double_k*p1)/(double_k+1.0);
         p1  = p2;
         p2  = p3;
@@ -31,26 +31,26 @@ float gll_library::pnleg(const float z, const int n){
     return p3;
 }
 
-float gll_library::pnglj(const float z, const int n){
+double gll_library::pnglj(const double z, const int n){
 
     if (n==0) throw std::invalid_argument("value of n > 0");
 
-    float glj_value;
+    double glj_value;
     
     if (std::abs(z+1.0) > 1e-9){
         glj_value = (gll_library::pnleg(z,n)+gll_library::pnleg(z,n+1))/(1.0+z);
     } else {
-        glj_value = (static_cast<float>(n)+1.0)*std::pow(-1.0,n);
+        glj_value = (static_cast<double>(n)+1.0)*std::pow(-1.0,n);
     }
 
     return glj_value;
 }
 
-float gll_library::pndleg(const float z, const int n){
+double gll_library::pndleg(const double z, const int n){
 
     if (n==0) throw std::invalid_argument("value of n > 0");
 
-    float p1,p2,p1d,p2d,p3,p3d,double_k;
+    double p1,p2,p1d,p2d,p3,p3d,double_k;
 
     p1 = 1.0;
     p2 = z;
@@ -59,7 +59,7 @@ float gll_library::pndleg(const float z, const int n){
     p3d = 1.0;
 
     for (int k = 1; k < n; k++){
-        double_k = static_cast<float>(k);
+        double_k = static_cast<double>(k);
         p3 = ((2.0*double_k+1.0)*z*p2 - double_k*p1)/(double_k+1.0);
         p3d = ((2.0*double_k+1.0)*p2 + (2.0*double_k+1.0)*z*p2d 
                 - double_k*p1d)/(double_k+1.0);
@@ -72,11 +72,11 @@ float gll_library::pndleg(const float z, const int n){
     return p3d;
 }
 
-float gll_library::pndglj(const float z, const int n){
+double gll_library::pndglj(const double z, const int n){
 
     if (n==0) throw std::invalid_argument("value of n > 0");
 
-    float glj_deriv;
+    double glj_deriv;
 
     if (std::abs(z+1.0) > 1e-9){
         glj_deriv = (gll_library::pndleg(z,n)+gll_library::pndleg(z,n+1))/(1.0+z)
